Started with functional tests for tiling but encountered a few issues
