+++ data/config/adaguc.tiling.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<Configuration>
+
+  <TempDir value="{ADAGUC_TMP}"/>
+  <Path value="{ADAGUC_PATH}/data"/>
+  <DataBase dbtype="sqlite" parameters="{ADAGUC_TMP}/adaguc.autoresource.db"/>
+  
+  <!-- WMS settings -->
+  <WMS>
+    <Title>ADAGUC WMS Service</Title>
+    <Abstract>This service demonstrates how the ADAGUC server can be used to create OGC services.</Abstract>
+    <RootLayer>
+      <Title>Root Layer Title</Title>
+      <Abstract>Root Layer Abstract</Abstract>
+    </RootLayer>
+    <!-- PNG 8 bit requests are forced to 32 bit: better for antialiasing and combining layers with many colors -->
+    <WMSFormat name="image/png" format="image/png32"/> 
+  </WMS>
+  
+  
+  <!-- WMS -->
+  <Include location="{ADAGUC_PATH}/data/config/includes/WMS.include.xml"/>
+
+  <!-- WCS settings -->
+  <WCS>
+    <Title>ADAGUC WCS Service</Title>
+    <Label>wcsLabel</Label>
+  </WCS>
+
+  <!-- WCSFormat -->
+  <Include location="{ADAGUC_PATH}/data/config/includes/WCSFormat.include.xml"/>
+
+  <!-- Projections -->
+  <Include location="{ADAGUC_PATH}/data/config/includes/Projection.include.xml"/>
+           
+  <!--Legends -->
+  <Include location="{ADAGUC_PATH}/data/config/includes/DefaultStyles.include.xml"/>
+
+  <AutoResource enableautoopendap="true" enablelocalfile="true" enablecache="false">
+    <Dir basedir="{ADAGUC_PATH}data/datasets" prefix="{ADAGUC_PATH}data/datasets/"/>
+    <ImageText attribute="source"></ImageText>
+  </AutoResource>
+
+  <Include location="{ADAGUC_PATH}/data/config/includes/DefaultLayers.include.xml"/>
+  
+  <OpenDAP enabled="true" path="opendap"/>
+  
+  <Layer type="database">
+    <Name>testdata</Name>
+    <Title>testdata</Title>
+    <FilePath>{ADAGUC_PATH}/data/datasets/testdata.nc</FilePath>
+    <Variable>testdata</Variable>
+    <RenderMethod>nearest</RenderMethod>
+    <TileSettings tilewidthpx="32" 
+                  tileheightpx="32" 
+                  tilecellsizex=".5"
+                  tilecellsizey=".5"
+                  left="-16" right="48"  bottom="36" top="68"
+                  tileprojection="+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
+                  minlevel="1"
+                  maxlevel="6"
+                  tilemode="nearest"
+                  debug="true"
+                  threads="1" 
+                  tilepath="{ADAGUC_TMP}/tiling"/>
+   </Layer>
+   
+    <Layer type="database">
+    <Name>testdatant</Name>
+    <Title>testdatant</Title>
+    <FilePath>{ADAGUC_PATH}/data/datasets/testdata.nc</FilePath>
+    <Variable>testdata</Variable>
+    <RenderMethod>nearest</RenderMethod>
+    <DataBaseTable>myown</DataBaseTable>
+  <!--  <TileSettings tilewidthpx="32" 
+                  tileheightpx="32" 
+                  tilecellsizex=".5"
+                  tilecellsizey=".5"
+                  left="-16" right="48"  bottom="36" top="68"
+                  tileprojection="+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
+                  minlevel="1"
+                  maxlevel="6"
+                  tilemode="nearest"
+                  debug="true"
+                  threads="1" 
+                  tilepath="{ADAGUC_TMP}/tiling"/>-->
+   </Layer>
+  <!-- End of configuration /-->
+</Configuration>
